// src/design-system/tokens/build/react-native/tokens.d.ts
export const fontGlobalSans: string;
export const fontGlobalSansAlt: string;
export const fontGlobalMono: string;
export const fontGlobalSansExpressive: string;
export const fontSize100: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize150: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize200: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize250: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize300: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize325: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize350: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize400: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize450: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize500: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize550: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize600: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize650: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontSize700: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const fontLineHeight300: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLineHeight350: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLineHeight400: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLineHeight450: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLineHeight500: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLineHeight550: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLineHeight600: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLineHeight650: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const fontWeight400: string;
export const fontWeight500: string;
export const fontWeight600: string;
export const fontWeight700: string;
export const fontLetterSpacing100: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLetterSpacing200: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLetterSpacing300: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
/** Auto-added: normal letter spacing */
export const fontLetterSpacing400: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLetterSpacing500: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLetterSpacing600: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const fontLetterSpacing700: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};